refactor: Reorganize NativeUi library with component-based folder structure
- Move each component into its own folder (EmptyState, IconBox, ListItem, etc.)
- Split services into NativeNavbar and NativeNavigation folders
- Split combined enum files into individual files (NuiIconBoxSize.cs, NuiIconBoxShape.cs, etc.)
- Make NativeNavbarManager and NativeNavigationManager internal
- Add DependencyInjection.cs with AddNativeUi() extension method
- Update namespaces to follow folder structure (EcoData.NativeUi.Components.*)
- Update _Imports.razor files with new namespaces
- Remove @using EcoData.NativeUi.Services from consuming files
