+++ src/Apps/EcoPortal/EcoPortal.Client/Features/HealthAlerts/Pages/AlertDetailPage.razor
@@ -1,5 +1,4 @@
 @page "/alerts/{AlertId:guid}"
-@using EcoData.NativeUi.Services
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Sensors.Application.Client
 @using EcoPortal.Client.Components
