+++ src/Apps/EcoPortal/EcoPortal.Client/Features/HealthAlerts/Pages/HealthAlertsPage.razor
@@ -1,6 +1,5 @@
 @page "/alerts"
 @using BlazorStaticNavigation
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Features.HealthAlerts.Components
 @inject INativeNavbarManager Navbar
