+++ src/Apps/EcoPortal/EcoPortal.Client/Features/OrganizationMembers/Pages/OrganizationTeamPage.razor
@@ -7,7 +7,6 @@
 @using EcoData.Organization.Contracts.Dtos
 @using EcoData.Organization.Contracts.Parameters
 @using EcoData.Organization.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Features.OrganizationMembers.Components
 @using EcoPortal.Client.Features.OrganizationMembers.Dialogs
