+++ src/Apps/EcoPortal/EcoPortal.Client/Features/OrganizationSensors/Pages/OrganizationSensorCreatePage.razor
@@ -6,7 +6,6 @@
 @using EcoData.Organization.Application.Client
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Sensors.Contracts.Requests
-@using EcoData.NativeUi.Services
 @using EcoData.Sensors.Application.Client
 @using BlazingSingularity.Commands
 @using EcoPortal.Client.Features.OrganizationSensors.Dialogs
