+++ src/Apps/EcoPortal/EcoPortal.Client/Features/OrganizationSensors/Pages/OrganizationSensorsPage.razor
@@ -5,7 +5,6 @@
 @using EcoData.Organization.Contracts.Dtos
 @using EcoData.Organization.Application.Client
 @using EcoData.Sensors.Contracts.Dtos
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Services
 @using EcoPortal.Client.Features.Sensors.Components
