+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Organizations/Pages/DiscoverOrganizationsPage.razor
@@ -5,7 +5,6 @@
 @using EcoData.Common.Pagination.Blazor
 @using EcoPortal.Client.Features.Organizations.Components
 @using BlazingSingularity.Signals
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @inject IOrganizationHttpClient OrganizationClient
 @inject INativeNavigationManager Navigation
