+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Organizations/Pages/MyOrganizationsPage.razor
@@ -1,7 +1,6 @@
 @page "/organizations/my"
 @using EcoData.Organization.Contracts.Dtos
 @using EcoData.Organization.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Features.Organizations.Components
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Services
