+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Organizations/Pages/OrganizationDetailsPage.razor
@@ -6,7 +6,6 @@
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Sensors.Contracts.Parameters
 @using EcoData.Sensors.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Features.Organizations.Services
 @using EcoPortal.Client.Services
 @using EcoPortal.Client.Components
