+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Organizations/Pages/OrganizationEditPage.razor
@@ -4,7 +4,6 @@
 @attribute [OrganizationPermission(Permissions.Organization.Update)]
 @using EcoData.Organization.Contracts.Dtos
 @using EcoData.Organization.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Services
 @using EcoPortal.Client.Features.Organizations.Components
