+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Organizations/Pages/OrganizationMapPage.razor
@@ -3,7 +3,6 @@
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Sensors.Contracts.Parameters
 @using EcoData.Sensors.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @implements IAsyncDisposable
 @inject IJSRuntime JS
