+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Sensors/Pages/SensorConfigurePage.razor
@@ -5,7 +5,6 @@
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Locations.Application.Client
 @using EcoData.Locations.Contracts.Dtos
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @using EcoPortal.Client.Components
 @using BlazingSingularity.Commands
