+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Sensors/Pages/SensorDetailPages/Components/SensorReadingsPreview.razor
@@ -1,7 +1,6 @@
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoData.Sensors.Contracts.Parameters
 @using EcoData.Sensors.Application.Client
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Features.Sensors.Components
 @inject ISensorReadingHttpClient ReadingClient
 @inject INativeNavigationManager Navigation
