+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Sensors/Pages/SensorDetailPages/SensorDetailAuthenticated.razor
@@ -2,7 +2,6 @@
 @using EcoData.Sensors.Application.Client
 @using EcoData.Locations.Application.Client
 @using EcoData.Locations.Contracts.Dtos
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Features.Organizations.Authorization
 @using EcoPortal.Client.Features.Sensors.Pages.SensorDetailPages.Components
