+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Sensors/Pages/SensorEditPage.razor
@@ -3,7 +3,6 @@
 @attribute [Authorize]
 @using EcoData.Sensors.Application.Client
 @using EcoData.Sensors.Contracts.Dtos
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Features.Sensors.Components
 @using EcoPortal.Client.Services
 @using EcoPortal.Client.Components
