+++ src/Apps/EcoPortal/EcoPortal.Client/Features/Sensors/Pages/SensorsPage.razor
@@ -1,5 +1,4 @@
 @page "/sensors"
-@using EcoData.NativeUi.Services
 @using EcoData.Sensors.Contracts.Dtos
 @using EcoPortal.Client.Features.Sensors.Components
 @inject INativeNavigationManager Navigation
