+++ src/Apps/EcoPortal/EcoPortal.Client/Layout/MainLayout.razor
@@ -1,7 +1,6 @@
 @inherits LayoutComponentBase
 @using BlazingSingularity
 @using BlazingSingularity.Commands
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Components
 @using EcoPortal.Client.Pages
 @using EcoPortal.Client.Services
