+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/AccountPage.razor
@@ -1,7 +1,6 @@
 @page "/account"
 @using BlazingSingularity
 @using BlazingSingularity.Commands
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @inject AuthStateService AuthState
 @inject ITabNavigationService Tabs
