+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/GettingStartedPage.razor
@@ -1,6 +1,5 @@
 @page "/getting-started"
 @using BlazorStaticNavigation
-@using EcoData.NativeUi.Services
 @inject INativeNavbarManager Navbar
 
 <PageTitle>Getting Started - EcoData</PageTitle>
