+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/LoginPage.razor
@@ -1,7 +1,6 @@
 @page "/login"
 @layout AuthLayout
 @using EcoPortal.Client.Layout
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @using EcoData.Identity.Contracts.Requests
 @using BlazingSingularity
