+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/MonitorPage.razor
@@ -1,5 +1,4 @@
 @page "/monitor"
-@using EcoData.NativeUi.Services
 @inject INativeNavigationManager Navigation
 @inject INativeNavbarManager Navbar
