+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/OrgsPage.razor
@@ -1,5 +1,4 @@
 @page "/orgs"
-@using EcoData.NativeUi.Services
 @using EcoPortal.Client.Services
 @inject INativeNavigationManager Navigation
 @inject INativeNavbarManager Navbar
