+++ src/Apps/EcoPortal/EcoPortal.Client/Pages/RegisterPage.razor
@@ -5,7 +5,6 @@
 @using EcoData.Identity.Contracts.Requests
 @using BlazingSingularity
 @using BlazingSingularity.Commands
-@using EcoData.NativeUi.Services
 @inject IAuthHttpClient AuthClient
 @inject INativeNavigationManager Navigation
