+++ src/Apps/EcoPortal/EcoPortal.Client/Program.cs
@@ -1,6 +1,6 @@
 using EcoData.Identity.Application.Client;
 using EcoData.Locations.Application.Client;
-using EcoData.NativeUi.Services;
+using EcoData.NativeUi;
 using EcoData.Organization.Application.Client;
 using EcoData.Sensors.Application.Client;
 using EcoPortal.Client.Authorization;
@@ -34,8 +34,7 @@ builder.Services.AddSingleton<IOrganizationCacheService, OrganizationCacheServic
 builder.Services.AddScoped<ILeafletMapService, LeafletMapService>();
 
 // Navigation services
-builder.Services.AddScoped<INativeNavigationManager, NativeNavigationManager>();
-builder.Services.AddScoped<INativeNavbarManager, NativeNavbarManager>();
+builder.Services.AddNativeUi();
 builder.Services.AddScoped<ITabNavigationService, TabNavigationService>();
 
 builder.Services.AddScoped<AuthStateService>();
