+++ src/Apps/EcoPortal/EcoPortal.Client/_Imports.razor
@@ -11,7 +11,14 @@
 @using EcoPortal.Client.Layout
 @using EcoPortal.Client.Components
 @using EcoData.NativeUi
-@using EcoData.NativeUi.Services
+@using EcoData.NativeUi.Components.EmptyState
+@using EcoData.NativeUi.Components.IconBox
+@using EcoData.NativeUi.Components.ListItem
+@using EcoData.NativeUi.Components.NativeNavbar
+@using EcoData.NativeUi.Components.NativeNavigation
+@using EcoData.NativeUi.Components.PageLayout
+@using EcoData.NativeUi.Components.SectionHeader
+@using EcoData.NativeUi.Components.StatusBadge
 @using EcoPortal.Client.Features.Organizations.Components
 @using EcoPortal.Client.Features.Sensors.Components
 @using EcoPortal.Client.Features.GettingStarted.Components
