+++ src/Features/Common/EcoData.NativeUi/Components/EmptyState/NuiEmptyState.razor
@@ -1,4 +1,4 @@
-@namespace EcoData.NativeUi
+@namespace EcoData.NativeUi.Components.EmptyState
 
 <div class="@GetContainerClass()">
     @if (!string.IsNullOrEmpty(Icon))
