+++ src/Features/Common/EcoData.NativeUi/Components/EmptyState/NuiEmptyState.razor.css
