+++ src/Features/Common/EcoData.NativeUi/Components/IconBox/NuiIconBox.razor
@@ -1,4 +1,4 @@
-@namespace EcoData.NativeUi
+@namespace EcoData.NativeUi.Components.IconBox
 
 <div class="@GetContainerClass()">
     <MudIcon Icon="@Icon" Size="@GetIconSize()" />
