+++ src/Features/Common/EcoData.NativeUi/Components/IconBox/NuiIconBoxShape.cs
@@ -1,0 +1,8 @@
+namespace EcoData.NativeUi.Components.IconBox;
+
+public enum NuiIconBoxShape
+{
+    Rounded,
+    Circle,
+    Square
+}
