+++ src/Features/Common/EcoData.NativeUi/Components/IconBox/NuiIconBoxSize.cs
@@ -1,0 +1,9 @@
+namespace EcoData.NativeUi.Components.IconBox;
+
+public enum NuiIconBoxSize
+{
+    Small,
+    Medium,
+    Large,
+    ExtraLarge
+}
