+++ src/Features/Common/EcoData.NativeUi/Components/ListItem/NuiListItem.razor
@@ -1,4 +1,4 @@
-@namespace EcoData.NativeUi
+@namespace EcoData.NativeUi.Components.ListItem
 
 <div class="@GetContainerClass()" @onclick="HandleClick">
     @if (!string.IsNullOrEmpty(Icon))
