+++ src/Features/Common/EcoData.NativeUi/Components/NativeNavbar/NavbarState.cs
@@ -1,0 +1,6 @@
+namespace EcoData.NativeUi.Components.NativeNavbar;
+
+/// <summary>
+/// Represents the complete navbar state including title and actions.
+/// </summary>
+public sealed record NavbarState(string? Title, IReadOnlyList<NavbarAction> Actions);
