+++ src/Features/Common/EcoData.NativeUi/Components/NativeNavigation/NavigationState.cs
@@ -1,19 +1,4 @@
-namespace EcoData.NativeUi.Services;
-
-/// <summary>
-/// Direction of the most recent navigation.
-/// </summary>
-public enum NavigationDirection
-{
-    /// <summary>No navigation has occurred yet.</summary>
-    None,
-
-    /// <summary>Navigated forward to a new page.</summary>
-    Forward,
-
-    /// <summary>Navigated back to a previous page.</summary>
-    Back
-}
+namespace EcoData.NativeUi.Components.NativeNavigation;
 
 /// <summary>
 /// Immutable snapshot of the current navigation state.
