--- src/Features/Common/EcoData.NativeUi/Components/NuiIconBoxEnums.cs
@@ -1,16 +1,0 @@
-namespace EcoData.NativeUi;
-
-public enum NuiIconBoxSize
-{
-    Small,
-    Medium,
-    Large,
-    ExtraLarge
-}
-
-public enum NuiIconBoxShape
-{
-    Rounded,
-    Circle,
-    Square
-}
