+++ src/Features/Common/EcoData.NativeUi/Components/PageLayout/NuiPageLayout.razor
@@ -1,5 +1,6 @@
-@namespace EcoData.NativeUi
-@using EcoData.NativeUi.Services
+@namespace EcoData.NativeUi.Components.PageLayout
+@using EcoData.NativeUi.Components.NativeNavbar
+@using EcoData.NativeUi.Components.NativeNavigation
 @inject INativeNavigationManager Navigation
 @inject INativeNavbarManager Navbar
