+++ src/Features/Common/EcoData.NativeUi/Components/SectionHeader/NuiSectionHeader.razor
@@ -1,4 +1,4 @@
-@namespace EcoData.NativeUi
+@namespace EcoData.NativeUi.Components.SectionHeader
 
 <div class="@GetContainerClass()">
     <div class="nui-section-header-content">
