+++ src/Features/Common/EcoData.NativeUi/Components/SectionHeader/NuiSectionHeader.razor.css
