+++ src/Features/Common/EcoData.NativeUi/Components/StatusBadge/NuiStatusBadge.razor
@@ -1,4 +1,4 @@
-@namespace EcoData.NativeUi
+@namespace EcoData.NativeUi.Components.StatusBadge
 
 <div class="@GetBadgeClass()">
     @if (ShowDot)
