+++ src/Features/Common/EcoData.NativeUi/Components/StatusBadge/NuiStatusBadge.razor.css
