+++ src/Features/Common/EcoData.NativeUi/Components/StatusBadge/NuiStatusSize.cs
@@ -1,0 +1,8 @@
+namespace EcoData.NativeUi.Components.StatusBadge;
+
+public enum NuiStatusSize
+{
+    Small,
+    Medium,
+    Large
+}
