+++ src/Features/Common/EcoData.NativeUi/_Imports.razor
@@ -1,3 +1,13 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Routing
 @using MudBlazor
+
+@* Component namespaces *@
+@using EcoData.NativeUi.Components.EmptyState
+@using EcoData.NativeUi.Components.IconBox
+@using EcoData.NativeUi.Components.ListItem
+@using EcoData.NativeUi.Components.NativeNavbar
+@using EcoData.NativeUi.Components.NativeNavigation
+@using EcoData.NativeUi.Components.PageLayout
+@using EcoData.NativeUi.Components.SectionHeader
+@using EcoData.NativeUi.Components.StatusBadge
